--- .all-contributorsrc
@@ -1,207 +1,0 @@
-{
-  "projectName": "50_days_50_web_project",
-  "projectOwner": "dhairyagothi",
-  "repoType": "github",
-  "repoHost": "https://github.com",
-  "files": [
-    "README.md"
-  ],
-  "imageSize": 85,
-  "commit": false,
-  "commitConvention": "angular",
-  "contributors": [
-    {
-      "login": "dhairyagothi",
-      "name": "Dhairya Gothi",
-      "avatar_url": "https://avatars.githubusercontent.com/u/142989448?v=4",
-      "profile": "https://github.com/dhairyagothi",
-      "contributions": [
-        "projectManagement",
-        "code",
-        "doc"
-      ]
-    },
-    {
-      "login": "deepesh611",
-      "name": "Deepesh Patil",
-      "avatar_url": "https://avatars.githubusercontent.com/u/123585104?v=4",
-      "profile": "https://github.com/deepesh611",
-      "contributions": [
-        "doc"
-      ]
-    },
-    {
-      "login": "Yaser-123",
-      "name": "T Mohamed Yaser",
-      "avatar_url": "https://avatars.githubusercontent.com/u/140265777?v=4",
-      "profile": "https://github.com/Yaser-123",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "MonishRaman",
-      "name": "Monish.R",
-      "avatar_url": "https://avatars.githubusercontent.com/u/140420819?v=4",
-      "profile": "https://github.com/MonishRaman",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "atharva012004",
-      "name": "Atharva Shihurkar",
-      "avatar_url": "https://avatars.githubusercontent.com/u/136626840?v=4",
-      "profile": "https://github.com/atharva012004",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Vansh-kash2023",
-      "name": "Vansh Kashyap",
-      "avatar_url": "https://avatars.githubusercontent.com/u/145321719?v=4",
-      "profile": "https://github.com/Vansh-kash2023",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "externref",
-      "name": "sarthak",
-      "avatar_url": "https://avatars.githubusercontent.com/u/70657005?v=4",
-      "profile": "https://github.com/externref",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Kaustav2410",
-      "name": "Kaustav2410",
-      "avatar_url": "https://avatars.githubusercontent.com/u/94297897?v=4",
-      "profile": "https://github.com/Kaustav2410",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "SuhainaFathimaM",
-      "name": "Suhaina Fathima M",
-      "avatar_url": "https://avatars.githubusercontent.com/u/153417476?v=4",
-      "profile": "https://github.com/SuhainaFathimaM",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "kom-senapati",
-      "name": "K Om Senapati ",
-      "avatar_url": "https://avatars.githubusercontent.com/u/92045934?v=4",
-      "profile": "https://github.com/kom-senapati",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "goutamiks",
-      "name": "Goutami K S",
-      "avatar_url": "https://avatars.githubusercontent.com/u/101326044?v=4",
-      "profile": "https://github.com/goutamiks",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "AnantAgarwaL11",
-      "name": "Anant Agarwal",
-      "avatar_url": "https://avatars.githubusercontent.com/u/115874061?v=4",
-      "profile": "https://github.com/AnantAgarwaL11",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Krishna100604",
-      "name": "Prajapati Shrikrushna R.",
-      "avatar_url": "https://avatars.githubusercontent.com/u/110299021?v=4",
-      "profile": "https://github.com/Krishna100604",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "deepak-d-r4113",
-      "name": "deepak-d-r4113",
-      "avatar_url": "https://avatars.githubusercontent.com/u/123359139?v=4",
-      "profile": "https://github.com/deepak-d-r4113",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Ayushswirlon",
-      "name": "Ayushswirlon",
-      "avatar_url": "https://avatars.githubusercontent.com/u/136717301?v=4",
-      "profile": "https://github.com/Ayushswirlon",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "shravn-10",
-      "name": "Sai Shravan M",
-      "avatar_url": "https://avatars.githubusercontent.com/u/109055682?v=4",
-      "profile": "https://github.com/shravn-10",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Madhu0-2",
-      "name": "Madhu0-2",
-      "avatar_url": "https://avatars.githubusercontent.com/u/126780854?v=4",
-      "profile": "https://github.com/Madhu0-2",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "nikitha2611",
-      "name": "Nikitha Koppuram",
-      "avatar_url": "https://avatars.githubusercontent.com/u/152257119?v=4",
-      "profile": "https://github.com/nikitha2611",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Shivakumarjagadam",
-      "name": "Shivakumarjagadam",
-      "avatar_url": "https://avatars.githubusercontent.com/u/154302898?v=4",
-      "profile": "https://github.com/Shivakumarjagadam",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Prashant-2024",
-      "name": "Prashant Solanki",
-      "avatar_url": "https://avatars.githubusercontent.com/u/140724156?v=4",
-      "profile": "https://github.com/Prashant-2024",
-      "contributions": [
-        "code"
-      ]
-    },
-    {
-      "login": "Umashankar08",
-      "name": "Umashankar08",
-      "avatar_url": "https://avatars.githubusercontent.com/u/143025358?v=4",
-      "profile": "https://github.com/Umashankar08",
-      "contributions": [
-        "code"
-      ]
-    }
-  ],
-  "contributorsPerLine": 7,
-  "linkToUsage": false
-}
