--- package.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "express": "^4.19.2",
-    "nodemon": "^3.1.3"
-  },
-  "scripts": {
-    "start": "nodemon index.js"
-  },
-  "devDependencies": {
-    "all-contributors-cli": "^6.26.1"
-  }
-}
