--- public/Login.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Login</title>
-    <link rel="stylesheet" href="login.css" />
-  </head>
-  <body>
-    <div class="container">
-      <div class="left-section">
-        <button
-          class="back-button"
-          onclick="window.location.href='index.html'"
-        ><img src="back1.png" width="50px" height="50px"/></button>
-        <h1 class="animated-heading">
-          50 Days 50 Web Projects!<span class="blinking-cursor">|</span>
-        </h1>
-        <p class="description">
-          Today, we present to you a sleek and modern login page. As part of our journey through 50 exciting web projects in 50 days, this login page exemplifies a perfect blend of style and functionality. Dive in to explore a user-friendly interface designed for secure and seamless access to your account. Enjoy the experience as we continue to innovate and create day by day!
-        </p>
-        <form id="loginForm" class="login-form">
-          <input
-            type="text"
-            id="username"
-            placeholder="Username"
-            class="input-field"
-          />
-          <input
-            type="password"
-            id="password"
-            placeholder="Password"
-            class="input-field"
-          />
-          <button type="submit" class="submit-button">Login</button>
-        </form>
-        <p class="already-signed-up">
-          Not signed up? <a href="./signup.html">Sign In</a>
-        </p>
-      </div>
-      <div class="right-section">
-        <img src="i1.jpg" alt="" class="transition-image" />
-        <img src="i2.jpg" alt= "" class="transition-image" />
-        <img src="im3.jpg" alt="" class="transition-image" />
-        <img src="im4.jpg"" alt="" class="transition-image" />
-        <img src="im5.jpg" alt="" class="transition-image" />
-      </div>
-    </div>
-
-    <script>
-      const images = document.querySelectorAll(".transition-image");
-      let currentImageIndex = 0;
-
-      function showNextImage() {
-        images[currentImageIndex].style.opacity = 0;
-        currentImageIndex = (currentImageIndex + 1) % images.length;
-        images[currentImageIndex].style.opacity = 1;
-      }
-
-      setInterval(showNextImage, 3000);
-
-      document.getElementById('loginForm').addEventListener('submit', function(event) {
-        event.preventDefault();
-
-        const username = document.getElementById('username').value;
-        const password = document.getElementById('password').value;
-
-        // Replace this with actual validation logic
-        if (username === 'test' && password === 'test') {
-          window.location.href = 'index.ejs';
-        } else {
-          alert('Invalid credentials. Please try again.');
-        }
-      });
-    </script>
-  </body>
-</html>
