--- public/signup.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Register</title>
-    <link rel="stylesheet" href="login.css"/>
-</head>
-<body>
-    <div class="container">
-        <div class="left-section">
-            <button class="back-button" onclick="window.location.href='index.html'">
-                <img src="back1.png" width="50px" height="50px"/>
-            </button>
-            <h1 class="animated-heading">50 Days 50 Web Projects!<span class="blinking-cursor">|</span></h1>
-            <p class="description">
-Thank you for joining us on this exciting journey of creativity and learning. Over the next 50 days, we will embark on a web development adventure, creating 50 unique web projects together. Whether you're a seasoned developer or just starting out, there's something here for everyone to explore, learn, and build.
-            </p>
-            <form id="signupForm" class="login-form">
-                <input type="text" id="name" placeholder="Name" class="input-field">
-                <input type="text" id="username" placeholder="Username" class="input-field">
-                <input type="email" id="email" placeholder="Email" class="input-field">
-                <input type="password" id="password" placeholder="Password" class="input-field">
-                <button type="submit" class="submit-button">Sign up</button>
-            </form>
-            <p class="already-signed-up">
-                already signed up? <a href="login.html">Login</a>
-            </p>
-        </div>
-        <div class="right-section">
-            <img src="i1.jpg" alt="" class="transition-image" />
-            <img src="i2.jpg" alt= "" class="transition-image" />
-            <img src="im3.jpg" alt="" class="transition-image" />
-            <img src="im4.jpg"" alt="" class="transition-image" />
-            <img src="im5.jpg" alt="" class="transition-image" />
-          </div>
-    </div>
-
-    
-
-    <script>
-        const images = document.querySelectorAll('.transition-image');
-        let currentImageIndex = 0;
-
-        function showNextImage() {
-            images[currentImageIndex].style.opacity = 0;
-            currentImageIndex = (currentImageIndex + 1) % images.length;
-            images[currentImageIndex].style.opacity = 1;
-        }
-
-        setInterval(showNextImage, 3000);
-
-        document.getElementById('signupForm').addEventListener('submit', function(event) {
-            event.preventDefault();
-
-            const name = document.getElementById('name').value;
-            const username = document.getElementById('username').value;
-            const email = document.getElementById('email').value;
-            const password = document.getElementById('password').value;
-
-            // Replace this with actual validation and signup logic
-            if (name && username && email && password) {
-                // Assume the registration is successful
-                window.location.href = 'index.html';
-            } else {
-                alert('Please fill in all fields.');
-            }
-        });
-    </script>
-</body>
-</html>
